Moves removeExistingKeys to LoadAllOperation
