+++ hazelcast/src/main/java/com/hazelcast/map/impl/recordstore/RecordStore.java
@@ -313,10 +313,9 @@ public interface RecordStore<R extends Record> extends LocalRecordStoreStats {
     /**
      * Loads all given keys from defined map store.
      *
-     * @param keys                  keys to be loaded.
-     * @param replaceExistingValues <code>true</code> if need to replace existing values otherwise <code>false</code>
+     * @param keys keys to be loaded.
      */
-    void loadAllFromStore(List<Data> keys, boolean replaceExistingValues);
+    void loadAllFromStore(List<Data> keys);
 
     void updateLoadStatus(boolean lastBatch, Throwable exception);
