+++ hazelcast/src/main/java/com/hazelcast/map/impl/recordstore/RecordStoreLoader.java
@@ -28,7 +28,7 @@ interface RecordStoreLoader {
 
     RecordStoreLoader EMPTY_LOADER = new RecordStoreLoader() {
         @Override
-        public Future loadValues(List<Data> keys, boolean replaceExistingValues) {
+        public Future loadValues(List<Data> keys) {
             return null;
         }
     };
@@ -36,9 +36,8 @@ interface RecordStoreLoader {
     /**
      * Loads all keys from defined map store.
      *
-     * @param keys                  keys to be loaded.
-     * @param replaceExistingValues <code>true</code> if need to replace existing values otherwise <code>false</code>
+     * @param keys keys to be loaded.
      * @return future for checking when loading is complete
      */
-    Future<?> loadValues(List<Data> keys, boolean replaceExistingValues);
+    Future<?> loadValues(List<Data> keys);
 }
